LayoutTests/fast/mediastream/play-newly-added-audio-track.html

<!DOCTYPE html>
<html>
    <head>
        <video id="video" autoplay controls></video>
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
        <script>
promise_test(async test => {
    const stream = await navigator.mediaDevices.getUserMedia({ video : true });
    test.add_cleanup(() => stream.getTracks().forEach(track => track.stop()));
    video.srcObject = stream;

    await video.play();

    video.srcObject.ontrackadded = () => {
        if (window.internals)
            assert_true(internals.shouldAudioTrackPlay(video.audioTracks[0]));
    };

    const stream2 = await navigator.mediaDevices.getUserMedia({ audio : true });
    test.add_cleanup(() => stream2.getTracks().forEach(track => track.stop()));
    video.srcObject.addTrack(stream2.getAudioTracks()[0]);
    await new Promise(resolve => setTimeout(resolve, 50));

    stream.getTracks().forEach(t => t.stop());
    stream2.getTracks().forEach(t => t.stop());
}, "Add an audio track while playing video");

promise_test(async test => {
    const stream = await navigator.mediaDevices.getUserMedia({ audio : true });
    test.add_cleanup(() => stream.getTracks().forEach(track => track.stop()));
    video.srcObject = stream.clone();

    await video.play();

    video.srcObject.ontrackadded = () => {
        if (window.internals)
            assert_true(internals.shouldAudioTrackPlay(video.audioTracks[1]));
    };

    video.srcObject.addTrack(stream.getAudioTracks()[0]);
    await new Promise(resolve => setTimeout(resolve, 50));

    stream.getTracks().forEach(t => t.stop());
}, "Add an audio track while playing audio");
        </script>
    </head>
</html>
