LayoutTests/platform/mac-wk2/fast/mediastream/play-newly-added-audio-track-expected.txt

CONSOLE MESSAGE: A capture MediaStreamTrack was destroyed without having been stopped explicitly
CONSOLE MESSAGE: A capture MediaStreamTrack was destroyed without having been stopped explicitly


PASS Add an audio track while playing video
PASS Add an audio track while playing audio
